" -FindAll -FindByid -HttpErrorResponse -New methods on BookService "
