+++ src/main/java/com/example/bookstore/BookstoreApplication.java
@@ -1,10 +1,7 @@
-package com.example.library;
+package com.example.bookstore;
 
-import com.example.library.repository.BookRepository;
-import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
 
 @SpringBootApplication
 public class BookstoreApplication {
@@ -12,7 +9,9 @@ public class BookstoreApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(BookstoreApplication.class, args);
 	}
-//
+
+
+	//
 //	@Bean
 //	CommandLineRunner runner(BookRepository bookRepository)
 //	{
