+++ src/main/java/com/example/bookstore/controller/BookStoreController.java
@@ -1,6 +1,6 @@
-package com.example.library.controller;
+package com.example.bookstore.controller;
 
-import com.example.library.service.BookStoreService;
+import com.example.bookstore.service.BookStoreService;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
