+++ src/main/java/com/example/bookstore/controller/HttpErrorResponse.java
@@ -1,0 +1,18 @@
+//HttpErrorResponse
+
+package com.example.bookstore.controller;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class HttpErrorResponse {
+    private int status;
+    private String Error;
+    private LocalDateTime timestamp;
+}
