+++ src/main/java/com/example/bookstore/controller/response/BookResponse.java
@@ -1,26 +1,26 @@
-package com.example.library.controller.response;
+package com.example.bookstore.controller.response;
 
 
-import com.example.library.model.BookStore;
-import com.example.library.model.BookType;
+import com.example.bookstore.model.BookStore;
+import com.example.bookstore.model.BookType;
 import lombok.*;
-
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
+import org.springframework.data.mongodb.core.mapping.Document;
 
 @AllArgsConstructor
 @Builder
 @NoArgsConstructor
 @ToString
-@Getter
-@Setter
+@Document
+@Data
 
 public class BookResponse
 {
+    //private String id;
     private String bookTitle;
     private String bookAuthor;
-    @Enumerated(EnumType.STRING)
     private BookType bookType;
     private BookStore bookStoreName;
     private BookStore bookStoreCity;
+
+
 }
