+++ src/main/java/com/example/bookstore/exception/BookNotFound.java
@@ -1,0 +1,15 @@
+
+//BookNotFound
+package com.example.bookstore.exception;
+
+public class BookNotFound extends RuntimeException {
+
+    public BookNotFound() {
+        super("Product not found.");
+    }
+
+    public BookNotFound(String message) {
+        super(message);
+    }
+}
+
