+++ src/main/java/com/example/bookstore/model/Book.java
@@ -1,0 +1,35 @@
+package com.example.bookstore.model;
+
+import lombok.*;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+import javax.persistence.*;
+import javax.validation.constraints.NotBlank;
+
+@Data
+@Document
+@Getter
+@Setter
+@ToString
+@AllArgsConstructor
+@Builder
+
+public class Book {
+
+    @Id
+    private String id;
+
+    @NotBlank(message = "Title name is mandatory")
+    @Column(name = "title")
+    private String bookTitle;
+
+    @Column(name = "author")
+    private String bookAuthor;
+
+    private BookType bookType; //VERIFY IF IT'S POSSIBLE
+
+    private BookStore bookStoreName;
+
+    private BookStore bookStoreCity;
+
+}
