+++ src/main/java/com/example/bookstore/model/BookType.java
@@ -1,0 +1,6 @@
+package com.example.bookstore.model;
+
+public enum BookType {
+    NEW,
+    USED
+}
