+++ src/main/java/com/example/bookstore/repository/BookStoreRepository.java
@@ -1,13 +1,11 @@
-package com.example.library.repository;
+package com.example.bookstore.repository;
 
 
-import com.example.library.model.BookStore;
+import com.example.bookstore.model.BookStore;
 import org.springframework.context.annotation.Primary;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 @Primary
@@ -15,6 +13,6 @@ import java.util.Optional;
 
 public interface BookStoreRepository extends MongoRepository<BookStore, String> {
 
-    Optional<BookStore> findByBookStoreName (String bookStoreName);
+    //Optional<BookStore> findByBookStoreName (String bookStoreName);
 
 }
