--- src/main/java/com/example/library/model/BookType.java
@@ -1,6 +1,0 @@
-package com.example.library.model;
-
-public enum BookType {
-    NEW,
-    USED
-}
