+++ src/test/java/com/example/bookstore/BookstoreApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.library;
+package com.example.bookstore;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
